[Scheduling] Extract a dialect interface for operator types.
This creates a new interface in the Scheduling library to allow
dialects to provide OperatorTypes used in scheduling problems. It
initially contains a single populateOperatorTypes method, which allows
the dialect to attach operator types and map them within a given
problem. Initially, this is used to replace the AffineToStaticLogic
use of Calyx operator types.